requirements.txt
tk==0.1.0
virtualenv==20.2.2
virtualenv-clone==0.5.4
numpy==1.19.5
matplotlib==3.5.0
matplotlib-inline==0.1.3
Pillow==8.4.0